[Mingu] Dev 복사 & 화로 로직 구현(예외처리 필요)

+++ Huck/Assets/01.UnityProject/Scripts/Interface/IDamageable.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class DamageMessage
 {
     public GameObject causer;
-    public float damageAmount;
+    public int damageAmount;
     public ItemData item;
-    public DamageMessage(GameObject causer, float damageAmount, ItemData item = default)
+    public DamageMessage(GameObject causer, int damageAmount, ItemData item = default)
     {
         this.causer = causer;
         this.damageAmount = damageAmount;
